simplify code by not using ManagedBlocker.
From my understanding ManagedBlocker is useful when witing on resources not managed by the pool.
When waiting on join, job stealing do occur so using a ManagedBlocker is not necessary.
However is the ManagedBlocker can have the side effect of increasing throughput due to additional threads.
But this is only because the default common pool has size (#processor - 1)
(also maybe. by reducing the probability that a thread could is waiting without having another job to steal).

By using a custom pool with #processor threads or a bit more, one can restore the 100% usage without ManagedBlocker.